🚚 Move leaderboard title to components
